fix(tests): refactor folders
