+++ test/views/pages/home/static_home_test.dart
@@ -7,10 +7,10 @@ import "package:proxima/views/home_content/feed/post_feed.dart";
 import "package:proxima/views/navigation/bottom_navigation_bar/navigation_bar_routes.dart";
 import "package:proxima/views/navigation/bottom_navigation_bar/navigation_bottom_bar.dart";
 import "package:proxima/views/pages/home/home_page.dart";
-
 import "package:proxima/views/pages/home/top_bar/app_top_bar.dart";
-import "../../viewmodels/mock_home_view_model.dart";
-import "../utils/mock_data/mock_posts.dart";
+
+import "../../../mocks/mock_posts.dart";
+import "../../../viewmodels/mock_home_view_model.dart";
 
 void main() {
   late ProviderScope homePageWidget;
